add npm start script to run local web-server for easier file viewing
Added a dev-dependency of `http-server`, and added a `start` script to launch  a server that can be accessed on `http://localhost:9009`.

This should make things a bit easier for many folks, instead of trying to enable the `file://` protocol in your browser of choice, setting up your own server, or needing to use an editor with built in connect.

To use, simply run `npm start` in your terminal, and then in your browser address bar, go to `http://localhost:9009`